+++ Libplanet.Net/Transports/TcpTransport.cs
@@ -279,7 +279,8 @@ namespace Libplanet.Net.Transports
                     // SocketException is thrown on .Net core when the given peer is invalid.
                     // To match with NetMQTransport implementation, convert to TimeoutException
                     // when failed to find the peer to send message.
-                    throw new TimeoutException($"Failed to send {message} {reqId} to {peer}.");
+                    throw new SendMessageFailException(
+                        $"Failed to send {message} {reqId} to {peer}.", peer, e);
                 }
 
                 var replies = new List<Message>();
@@ -379,6 +380,7 @@ namespace Libplanet.Net.Transports
                 return replies;
             }
             catch (Exception e) when (
+                e is SendMessageFailException ||
                 e is InvalidMagicCookieException ||
                 e is InvalidMessageSignatureException ||
                 e is InvalidMessageTimestampException ||
@@ -389,7 +391,11 @@ namespace Libplanet.Net.Transports
                     "Failed to send and receive replies from {Peer} for request " +
                     "{Message} {RequestId}.";
                 _logger.Error(e, errMsg, peer, message, reqId);
-                throw;
+                throw new CommunicationFailException(
+                    $"Failed to send and receive replies from {peer} for request {message}.",
+                    message.Type,
+                    peer,
+                    e);
             }
             finally
             {